- Simplify mfs_file_data_from_path a lot and rename it to mfs_realpath, which   have the objective of looking up the real file path, given the musicfs path.
Some thoughts:
Instead of creating a function for each operation we want to do, this change
adds a mfs_lookup_path which allocates a buffer with the real path of the file
we're looking for. This way, we can do the operations we want on the path
externally instead, which is a lot simpler when using it. This also gets rid of
the file_data data structure, and just use a pointer pointer where one is to
allocate the real path.